bot/requirements.txt
aiogram
requests
python_dotenv